+++ R/lifecycle-retired.R
@@ -1173,7 +1173,7 @@ is_bare_env <- function(x) {
 #' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}
 #'
 #' As of rlang 0.3.0, `env_bind_exprs()` and `env_bind_fns()` have
-#' been renamed to [env_bind_promise()] and [env_bind_active()] for
+#' been renamed to [env_bind_lazy()] and [env_bind_active()] for
 #' consistency.
 #'
 #' @inheritParams env_bind
@@ -1183,9 +1183,9 @@ is_bare_env <- function(x) {
 env_bind_exprs <- function(.env, ..., .eval_env = caller_env()) {
   signal_soft_deprecated(paste_line(
     "`env_bind_exprs()` is soft-deprecated as of rlang 0.3.0.",
-    "Please use `env_bind_promise()` instead."
+    "Please use `env_bind_lazy()` instead."
   ))
-  env_bind_promise(.env = .env, ..., .eval_env = .eval_env)
+  env_bind_lazy(.env = .env, ..., .eval_env = .eval_env)
 }
 #' @rdname env_bind_exprs
 #' @export