+++ NEWS.md
@@ -151,10 +151,10 @@
 * Quosures converted to functions with `as_function()` now support
   nested quosures.
 
-* `env_bind_promise()` and `env_bind_active()` now support quosures.
+* `env_bind_lazy()` and `env_bind_active()` now support quosures.
 
 * `env_bind_exprs()` and `env_bind_fns()` are soft-deprecated and
-  renamed to `env_bind_promise()` and `env_bind_active()` for clarity
+  renamed to `env_bind_lazy()` and `env_bind_active()` for clarity
   and consistency.
 
 * `env_bind()`, `env_bind_exprs()`, and `env_bind_fns()` now return